+++ lib/src/socket/socket.dart
@@ -4,14 +4,13 @@ import 'dart:async';
 import 'dart:collection';
 import 'dart:convert';
 import 'dart:io';
-import 'package:collection/collection.dart' show IterableExtension;
 import 'package:get_server/src/framework/get_core/get_core.dart';
 
 part 'socket_impl.dart';
 part 'socket_interface.dart';
 part 'socket_notifier.dart';
 
-extension FirstWhereExt<T> on List<T> {
+extension FirstWhereExt<T> on Iterable<T> {
   T? firstWhereOrNull(bool Function(T element) test) {
     for (var element in this) {
       if (test(element)) return element;